+++ su065d4380_interface/test/src/commander/test_info_commander.cpp
@@ -33,9 +33,11 @@ public:
   {
   }
 
-  MOCK_METHOD(int, getBytesAvailable, (), (const override));
-  MOCK_METHOD(int, readPort, (char *, const int), (const override));
-  MOCK_METHOD(int, writePort, (const char *, const int), (const override));
+  MOCK_METHOD(size_t, getBytesAvailable, (), (const override));
+  MOCK_METHOD(size_t, readPort, (char *, const size_t), (const override));
+  MOCK_METHOD(
+    size_t, writePort,
+    (const char *, const size_t), (const override));
 };
 
 using namespace std::chrono_literals;
